migration_pjt/usdms_origin/backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
psycopg2-binary
pandas
numpy
python-dotenv
apscheduler
requests
aiohttp
beautifulsoup4
lxml
pydantic-settings
tqdm
yfinance
